fixed wrong types import; separated types and data for creation script; removed unused types
